Support PTL to run non-root commands from other hosts using IP and port
